+++ run_disney_dj.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+
+export PYGAME_HIDE_SUPPORT_PROMPT='hide'
+
+./disc_jockey.py \
+  -n 16 \
+  -d $HOME/Desktop/Disney \
+  -r 1.0